requirements.txt
Django==4.2
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
drf_yasg==1.21.7
pytest==8.0.0
pytest-django==4.8.0
pillow==10.2.0
celery==5.3.6
redis==5.0.1
django-storages==1.14.2
boto3==1.34.32
djangorestframework-simplejwt==5.3.1
django-filter==23.5
pytest-factoryboy==2.6.0